requirements.txt
fastapi
uvicorn
groq
python-docx
python-dotenv
docx2pdf
gspread
google-auth
